requirements.txt
flask
wordninja
pyspellchecker